KQueue
The KQueue simulate that a system. In this system, customers arrive as usual from the outside. Now there are K servers available. A customer can go to any of the K servers for service. Thus, all K of them have to be busy serving for a customer to need to wait.
You could directly run KQueue class and then input a k value which is the number of servers. you also can set the arrivalRate value in the main method. 
The RadomTool class is provided by Professor Simha.